CPOO TP6 Ex3 - Projet logique

+++ Licence_Informatique/L3/S5/CPOO/TP/TP6/Additionneur.java
@@ -9,6 +9,6 @@ public class Additionneur implements UnaryOperator<Integer> {
 
     @Override
     public Integer apply(Integer integer) {
-        return integer + x;
+        return (integer.intValue()+ x);
     }
 }
